third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-srcset-react-to-media-changes-when-image-not-changed-expected.txt

PASS img.currentSrc.indexOf("blue_rect") is not -1
PASS img.width is 70
PASS img.currentSrc.indexOf("blue_rect") is not -1
PASS img.width is 75
PASS successfullyParsed is true

TEST COMPLETE
